make "getVelocity" a function
when the local versions of the various functions were introduced, a
macro was used so that old code would function as before, even though
the work is now performed either in getVelocitySoap or getVelocityLocal.
however, that means that python will not see the function named
"getVelocity" unless CUTOUT_SUPPORT was on for the compilation (which
will not always be true).
therefore I changed the code so that the function is named properly, and
we don't have to explain to people that, for python, the relevant
function has "Soap" at the end for no obvious reason.

git-svn-id: svn+ssh://pha/home/SVN/trunk/turbulence/turblib@10992 743f73a2-0892-4f1a-8574-ff7a07c3af4f